Add framework for event system
An event is a simple class broadly categorized by its EventCategory bitfield
and uniquely defined by its EventType enum.

Events are handled via the EventDispatcher class. The dispatcher will
check to see if the event in the dispatch call is of the same EventType
as its member variable event. If so, it will invoke an EventFn which is
a function that returns a bool and takes an Event by reference.

End of video: Event System