Add example of EventDispatcher class in Application
Whenever an event occurs, it is sent to the OnEvent function in
Application.cpp. The function then constructs an EventDispatcher with
the given event. The dispatch call will now only dispatch a given
function if it matches the provided event type.

This example handles the WindowCloseEvent which is triggered when the
GLFW window's X button is pressed. This triggers the Application to set
m_Running to false and causes the loop in the Run function to end and
the main thread to exit and our GLFW to close.

Comments were added to the EventDispatcher class for my own
understanding.

End of video: Window Events