Set MD for dynamic linking in premake
This appears to be a temporary solution to force our SandBox to
dynamically link against the Hazel lib. The problem with it is that it
does not set the correct field in VS and only forces dynamic linking via
a command line argument addition. I imagine a more permanent solution
has been found at this point, but I will stick with the intent of the
original video.