Add Layer and LayerStack
A layer is an element that is intended to be drawn.
A layer stack is how Hazel will manage the layers of an application.

The layer stack groups layers into two categories: layers and overlays.
Layers are drawn first and overlays are drawn last. However, when an event
is to be handled, the overlays are queried first and then the layers.