Fix event bugs
MouseButtonReleasedEvent was labeled as a PressedEvent in the
WindowsWindow class. This caused the release event to never be properly
set.

Similarly, MouseScrolledEvent was returning a MouseMovedType in its
GetStaticType() call. This made it so that event dispatchers of Scrolled
and Moved typed would not operate properly.

These bugs were found while trying to integrate ImGui into the
WindowsWindow and were an absolute nightmare to track.