Add and implement Input interface for polling
The Input class defines an interface for polling particular inputs on
any platform. Namely cursor position, key pressed status and mouse press
status. The Window class was updated to return a void pointer of its
native window to allow for the Input implementation class to access its
fields for polling.

WindowsInput is an implementation of Input that returns the values
specified in the Input interface via GLFW. If the platform were
different, or if a Windows implementation decided not to use GLFW, Input
is generic enough to allow for different implementations.

The Application run loop was updated to demonstrate input polling
working on mouse position.

End of video: Input Polling