Add vertex buffer layouts
Vertex buffer layouts describe the specifics of a vertex binary loaded
by the graphics driver. Given a binary comprised of a vec3 of position
verticies and a vec4 of colors, the engine will be responsible for
calculating the following values:

Stride: The size of a single combined element in the binary
(sizeof(vec3) + sizeof(vec4))

Offset: The start of each type in the vertex binary. Starts at 0 and
increases based on size of added type. vec3 starts at byte 0 and vec4
starts at byte sizeof(vec3). If an additional type were added, its
offset would be sizeof(vec3) + sizeof(vec4).

The implementation currently only for OpenGL and exists in the
Application class, but will be abstracted and moved to the OpenGLBuffer
class.

End of video: Vertex Buffer Layouts