Add VertexArray
Vertex arrays keep a reference to a vertex buffer and an index buffer.
In OpenGL, this allows for a reduction in the number of duplicate binds
when rendering multiple vertex buffers. However, this construct does not
appear to be used in DirectX or Vulkan and does not seem to be well
liked.

End of video: Vertex Arrays