Add high level Renderer class and associated files
The Renderer class is responsible for assembling the necessary pieces
for to render a scene. Currently, that means clearing the current window
and adding vertex arrays to render. In the future, it will be able to
take many types by overloading the renderer's Submit() function.

End of video: Renderer Flow and Submission