Add orthographic camera
A camera can now be set by the renderer's BeginScene() call. The
renderer will use the camera's ViewProjectionMatrix to draw the scene in
the position of the camera.

The orthographic camera's ViewProjectionMatrix is used to render 2D
scenes whereas the yet to be added projection camera's
ViewProjectionMatrix is used to render 3D scenes.