Abstract Shader class from OpenGL implementation
The old Shader class was effectively an OpenGLShader class. This patch
abstracts away the OpenGL components into its own class (OpenGLShader)
and attempts to maintain the same API. Unfortunately, some OpenGL
specific calls are still public in the OpenGLShader class that are not
present in the new abstract Shader class. For the time being, some
shoddy downcasting is being performed as a temporary workaround that
will be replaced in the future.